+++ mobile/android/android-components/components/feature/contextmenu/src/test/java/mozilla/components/feature/contextmenu/ContextMenuFeatureTest.kt
@@ -7,8 +7,11 @@ package mozilla.components.feature.contextmenu
 import android.content.Context
 import android.support.v4.app.FragmentManager
 import android.support.v4.app.FragmentTransaction
+import android.view.HapticFeedbackConstants
+import android.view.View
 import mozilla.components.browser.session.Session
 import mozilla.components.browser.session.SessionManager
+import mozilla.components.concept.engine.EngineView
 import mozilla.components.concept.engine.HitResult
 import mozilla.components.support.base.observer.Consumable
 import mozilla.components.support.test.any
@@ -17,6 +20,7 @@ import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
+import org.mockito.Mockito.`when`
 import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
 import org.mockito.Mockito.verify
@@ -36,14 +40,21 @@ class ContextMenuFeatureTest {
 
         val fragmentManager = mockFragmentManager()
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            fragmentManager, sessionManager, ContextMenuCandidate.defaultCandidates(context, mock(), mock()))
+            fragmentManager,
+            sessionManager,
+            ContextMenuCandidate.defaultCandidates(context, mock(), mock()),
+            engineView)
+
         feature.start()
 
         val hitResult = HitResult.UNKNOWN("https://www.mozilla.org")
         session.hitResult = Consumable.from(hitResult)
 
         verify(fragmentManager).beginTransaction()
+        verify(view).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
@@ -53,8 +64,14 @@ class ContextMenuFeatureTest {
 
         val fragmentManager = mockFragmentManager()
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            fragmentManager, sessionManager, ContextMenuCandidate.defaultCandidates(context, mock(), mock()))
+            fragmentManager,
+            sessionManager,
+            ContextMenuCandidate.defaultCandidates(context, mock(), mock()),
+            engineView)
+
         feature.start()
         feature.stop()
 
@@ -62,6 +79,7 @@ class ContextMenuFeatureTest {
         session.hitResult = Consumable.from(hitResult)
 
         verify(fragmentManager, never()).beginTransaction()
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
@@ -78,12 +96,18 @@ class ContextMenuFeatureTest {
         val fragmentManager: FragmentManager = mock()
         doReturn(fragment).`when`(fragmentManager).findFragmentByTag(any())
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            fragmentManager, sessionManager, ContextMenuCandidate.defaultCandidates(context, mock(), mock()))
+            fragmentManager,
+            sessionManager,
+            ContextMenuCandidate.defaultCandidates(context, mock(), mock()),
+            engineView)
 
         feature.start()
 
         verify(fragment).feature = feature
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
@@ -103,13 +127,20 @@ class ContextMenuFeatureTest {
         doReturn(transaction).`when`(fragmentManager).beginTransaction()
         doReturn(transaction).`when`(transaction).remove(fragment)
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            fragmentManager, sessionManager, ContextMenuCandidate.defaultCandidates(context, mock(), mock()))
+            fragmentManager,
+            sessionManager,
+            ContextMenuCandidate.defaultCandidates(context, mock(), mock()),
+            engineView)
 
         feature.start()
 
         verify(fragmentManager).beginTransaction()
         verify(transaction).remove(fragment)
+
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     fun `Already existing fragment will be removed if session does not exist anymore`() {
@@ -125,13 +156,20 @@ class ContextMenuFeatureTest {
         doReturn(transaction).`when`(fragmentManager).beginTransaction()
         doReturn(transaction).`when`(transaction).remove(fragment)
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            fragmentManager, sessionManager, ContextMenuCandidate.defaultCandidates(context, mock(), mock()))
+            fragmentManager,
+            sessionManager,
+            ContextMenuCandidate.defaultCandidates(context, mock(), mock()),
+            engineView)
 
         feature.start()
 
         verify(fragmentManager).beginTransaction()
         verify(transaction).remove(fragment)
+
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
@@ -145,14 +183,17 @@ class ContextMenuFeatureTest {
             action = { _, _ -> Unit }
         )
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            fragmentManager, mock(), listOf(candidate))
+            fragmentManager, mock(), listOf(candidate), engineView)
 
         feature.onLongPress(
             Session("https://www.mozilla.org"),
             HitResult.UNKNOWN("https://www.mozilla.org"))
 
         verifyNoMoreInteractions(fragmentManager)
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
@@ -163,14 +204,21 @@ class ContextMenuFeatureTest {
             hitResult = Consumable.from(HitResult.UNKNOWN("https://www.mozilla.org"))
         }
 
+        val (engineView, view) = mockEngineView()
+
         val feature = ContextMenuFeature(
-            mock(), sessionManager, ContextMenuCandidate.defaultCandidates(context, mock(), mock()))
+            mock(),
+            sessionManager,
+            ContextMenuCandidate.defaultCandidates(context, mock(), mock()),
+            engineView)
 
         assertFalse(session.hitResult.isConsumed())
 
         feature.onMenuCancelled(session.id)
 
         assertTrue(session.hitResult.isConsumed())
+
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     @Test
@@ -189,7 +237,13 @@ class ContextMenuFeatureTest {
             showFor = { _, _ -> true },
             action = { _, _ -> actionInvoked = true })
 
-        val feature = ContextMenuFeature(mock(), sessionManager, listOf(candidate))
+        val (engineView, view) = mockEngineView()
+
+        val feature = ContextMenuFeature(
+            mock(),
+            sessionManager,
+            listOf(candidate),
+            engineView)
 
         assertFalse(session.hitResult.isConsumed())
         assertFalse(actionInvoked)
@@ -198,6 +252,8 @@ class ContextMenuFeatureTest {
 
         assertTrue(session.hitResult.isConsumed())
         assertTrue(actionInvoked)
+
+        verify(view, never()).performHapticFeedback(HapticFeedbackConstants.LONG_PRESS)
     }
 
     private fun mockFragmentManager(): FragmentManager {
@@ -207,5 +263,15 @@ class ContextMenuFeatureTest {
         doReturn(transaction).`when`(fragmentManager).beginTransaction()
 
         return fragmentManager
+    }
+
+    private fun mockEngineView(): Pair<EngineView, View> {
+        val actualView: View = mock()
+
+        val engineView = mock<EngineView>().also {
+            `when`(it.asView()).thenReturn(actualView)
+        }
+
+        return Pair(engineView, actualView)
     }
 }